moved contracts and kubix balance to web3 context
